agents/requirements.txt
# AgentOS Requirements - Updated 2025-12-13
# Reference: docs/detailed-implementation-plan.md Section 6

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# AI Agent Framework with Protocol Support
# https://docs.agno.com/how-to/install
# Agno with AG-UI and A2A protocol extras for DM-02
agno[agui,a2a]>=0.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Authentication
pyjwt==2.8.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
requests>=2.31.0
httpx>=0.27.0
respx>=0.20.2
slowapi>=0.1.9
limits>=3.7.0

# AG-UI Streaming (Server-Sent Events)
sse-starlette>=1.8.0

# A2A Performance (High-performance JSON)
orjson>=3.9.0

# RAG Knowledge Base
pgvector>=0.2.5
tiktoken>=0.5.0  # Token counting for embeddings

# Document Processing (for Agno readers)
pypdf>=3.0.0  # PDF reading
python-docx>=1.0.0  # DOCX reading
python-pptx>=0.6.0  # PPTX reading
beautifulsoup4>=4.12.0  # Web scraping
lxml>=5.0.0  # HTML/XML parsing

# BYOAI Key Encryption
cryptography>=41.0.0

# Model Context Protocol (MCP)
mcp>=0.1.0

# OpenTelemetry Observability (DM-09.1)
# Core tracing API and SDK
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0

# Auto-instrumentation for FastAPI, HTTPX, Redis
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-instrumentation-redis>=0.43b0
